pyproject.toml
[project]
name = "audit-agent"
version = "0.1.0"
description = "Risk-aware claim verification agent for auditing LLM-generated financial document outputs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "openai>=1.30",
    "pydantic>=2.7",
    "typer>=0.12",
    "rich>=13.7",
    "rapidfuzz>=3.9",
    "tenacity>=8.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "socksio>=1.0",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0",
    "matplotlib>=3.9",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
audit-agent = "audit_agent.cli:app"

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/audit_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
